+++ ofxsMultiPlane.cpp
@@ -1089,7 +1089,7 @@ static bool findBuiltInSelectedChannel(const std::string& selectedOptionID,
 
             if (channelOptionID == optionWithoutClipPrefix) {
 
-                int chanIndex = c;
+                int chanIndex = (int)c;
                 // If the hard-coded plane is the color plane, the channel may not exist actually in the available components,
                 // e.g: Alpha may be present in the choice but the components may be RGB
                 // In this case, return 1 instead for Alpha and 0 for any other channel.
